Add BSD-3-Clause-LBNL license
NOTE: The license is listed in the table of contents on the OSI website,
however, the link points to the 0BSD license.  There is no page for this
license on the OSI website - likely an error on the website itself.  The
name and text were takend from the SPDX data.

Signed-off-by: Gary O'Neall <gary@sourceauditor.com>