Adjust text scraping script to current OSI website
Repair the Python script for scraping the license texts from the OSI
website. It searches by one of the classes with which the main text
container element is marked in the HTML document; and these classes have
changed.

Also make the search more robust: a simple equality check is not
suitable because there may be several classes in the HTML ‘class’
attribute. The lxml library has a special function which covers exactly
that purpose.

While at it, switch from HTTP to HTTPS for accessing the web content.